# docs/swagger.yaml
definitions:
  models.User:
    properties:
      dob:
        type: string
      email:
        type: string
      fullname:
        type: string
      gender:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  request.CreateUserRequest:
    properties:
      dob:
        type: string
      email:
        type: string
      fullname:
        maxLength: 50
        minLength: 2
        type: string
      gender:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - fullname
    - gender
    - password
    - phone
    - username
    type: object
  request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.LogoutRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  response.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.UserResponse:
    properties:
      dob:
        type: string
      email:
        type: string
      fullname:
        type: string
      gender:
        type: string
      id:
        type: integer
      phone:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: duy@example.com
    name: Duy
    url: http://github.com/daoduy
  description: Đây là tài liệu API cho project học Golang
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Học Golang API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Cho phép người dùng đăng nhập bằng username và password
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Đăng nhập
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Thêm token vào danh sách vô hiệu hoá
      parameters:
      - description: Token cần vô hiệu hoá
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Đăng xuất người dùng
      tags:
      - Auth
  /users/:
    get:
      description: Trả về danh sách user trong hệ thống
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Lấy danh sách tất cả người dùng
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Tạo mới một người dùng trong hệ thống
      parameters:
      - description: Thông tin tạo người dùng
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Tạo mới người dùng
      tags:
      - User
  /users/{id}:
    get:
      description: Trả về thông tin chi tiết của người dùng theo ID
      parameters:
      - description: ID của người dùng
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserResponse'
      summary: Lấy thông tin người dùng theo ID
      tags:
      - User
swagger: "2.0"